+++ src/AppserverIo/Appserver/Core/Api/AppService.php
@@ -37,11 +37,18 @@ class AppService extends AbstractService
 {
 
     /**
-     * The unique XML configuration node name for a app node.
+     * The name of the default extractor we will look for in our configuration
      *
-     * @var string
+     * @var string DEFAULT_EXTRACTOR_NAME
      */
-    const NODE_NAME = 'application';
+    const DEFAULT_EXTRACTOR_NAME = 'phar';
+
+    /**
+     * The extractor instance to handle archive operations with
+     *
+     * @var \AppserverIo\Appserver\Core\Interfaces\ExtractorInterface $extractor
+     */
+    protected $extractor;
 
     /**
      * Creates a new app node for the passed application and attaches
@@ -53,17 +60,10 @@ class AppService extends AbstractService
      */
     public function newFromApplication(ApplicationInterface $application)
     {
-
-        // load the system configuration
-        $systemConfiguration = $this->getInitialContext()->getSystemConfiguration();
-
-        // create a new AppNode and initialize it with the values from this instance
+        // create a new AppNode and initialize it
         $appNode = new AppNode();
-        $appNode->setNodeName(AppService::NODE_NAME);
-        $appNode->setName($application->getName());
-        $appNode->setWebappPath($application->getWebappPath());
-        $appNode->setParentUuid($systemConfiguration->getUuid());
-        $appNode->setUuid($appNode->newUuid());
+        $appNode->initFromApplication($application);
+        $appNode->setParentUuid($this->getSystemConfiguration()->getUuid());
 
         // persist the AppNode instance
         $this->persist($appNode);
@@ -100,6 +100,44 @@ class AppService extends AbstractService
             }
         }
         return $appNodes;
+    }
+
+    /**
+     * Getter for this service's extractor
+     *
+     * @return \AppserverIo\Appserver\Core\Interfaces\ExtractorInterface|null
+     */
+    public function getExtractor()
+    {
+        // we will have a phar extractor by default
+        if (!isset($this->extractor)) {
+            $configuration = $this->getSystemConfiguration()->getExtractors();
+            if (isset($configuration[self::DEFAULT_EXTRACTOR_NAME])) {
+                // create a new extractor with the default configuration
+                $this->extractor = new PharExtractor($this->getInitialContext(), $configuration[self::DEFAULT_EXTRACTOR_NAME]);
+
+            } else {
+                $this->getInitialContext()->getSystemLogger()->warning(sprintf(
+                    'Did not find configuration for default extractor %s nor was an extractor injected.',
+                    self::DEFAULT_EXTRACTOR_NAME
+                ));
+                $this->extractor = null;
+            }
+        }
+
+        return $this->extractor;
+    }
+
+    /**
+     * Will inject a certain extractor to be used
+     *
+     * @param \AppserverIo\Appserver\Core\Interfaces\ExtractorInterface $extractor The extractor instance to inject
+     *
+     * @return null
+     */
+    public function injectExtractor(ExtractorInterface $extractor)
+    {
+        $this->extractor = $extractor;
     }
 
     /**
@@ -160,8 +198,8 @@ class AppService extends AbstractService
      */
     public function soak(\SplFileInfo $archive)
     {
-        $p = new PharExtractor($this->getInitialContext());
-        $p->soakArchive($archive);
+        $extractor = $this->getExtractor();
+        $extractor->soakArchive($archive);
     }
 
     /**
@@ -175,13 +213,13 @@ class AppService extends AbstractService
     public function deploy(NodeInterface $appNode)
     {
         // prepare file name
-        $fileName = $appNode->getName() . PharExtractor::EXTENSION_SUFFIX;
+        $extractor = $this->getExtractor();
+        $fileName = $appNode->getName() . $extractor->getExtensionSuffix();
 
         // load the file info
         $archive = new \SplFileInfo($this->getDeployDir() . DIRECTORY_SEPARATOR . $fileName);
 
-        // flag the archiv => deploy it with the next restart
-        $extractor = new PharExtractor($this->getInitialContext());
+        // flag the archive => deploy it with the next restart
         $extractor->flagArchive($archive, ExtractorInterface::FLAG_DODEPLOY);
     }
 
@@ -200,37 +238,14 @@ class AppService extends AbstractService
         // try to load the app node with the passe UUID
         if ($appNode = $this->load($uuid)) {
             // prepare file name
-            $extractor = new PharExtractor($this->getInitialContext());
+            $extractor = $this->getExtractor();
             $fileName = $appNode->getName() . $extractor->getExtensionSuffix();
 
             // load the file info
             $archive = new \SplFileInfo($this->getDeployDir() . DIRECTORY_SEPARATOR . $fileName);
 
-            // unflag the archiv => undeploy it with the next restart
-            $extractor = new PharExtractor($this->getInitialContext());
+            // un-flag the archiv => un-deploy it with the next restart
             $extractor->unflagArchive($archive);
         }
-    }
-
-    /**
-     * Returns an new app node instance.
-     *
-     * @param \AppserverIo\Psr\Application\ApplicationInterface $application The application to create a new app node instance from
-     *
-     * @return \AppserverIo\Appserver\Core\Api\Node\AppNode The app node representation of the application
-     */
-    protected function create(ApplicationInterface $application)
-    {
-
-        // create a new AppNode and initialize it with the values from this instance
-        $appNode = new AppNode();
-        $appNode->setNodeName(AppService::NODE_NAME);
-        $appNode->setUuid($appNode->newUuid());
-        $appNode->setName($application->getName());
-        $appNode->setWebappPath($application->getWebappPath());
-        $appNode->setDatasources($application->getDatasources());
-
-        // return the AppNode instance
-        return $appNode;
     }
 }