+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Bugfixes
 
+* Added missing dependency to appserver-io/lang
 * Fixed MQ memory leak because of missing job thread when handling messages
 * Fixed invalid namespace in QueueManager::createSenderForQueue() method
 * Remove unnecessary interfaces SenderInterface + ReceiverInterface